Shutdown correctly when requested via signal
Before this change, `forwarder` can hang while attempting to shutdown.

Because `multitail2` can block indefinitely, once the `receiver` is
down, `forwarder` must not try to `join()` the spool-reading thread
(i.e. `MetricForwarder`). Since it is a `daemon` thread, it will just
exit with the interpreter - and that's fine.

Added comments describing how this works, plus a few minor logging
changes.