Simplify backoff & make it exponential w/jitter
The existing backoff algorithm ends up sleeping as long as a
connectivity issue has been, and does a simple linear count; we can
envisage a "thundering herd" of agent clients all sending their data at
once, and delayed more than it has to be.

This approach (which is also easier to read) backs off exponentially
with a ceiling, and adds some initial randomness.